fix: render falsy number values
